requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
pydantic-settings==2.0.3
Pillow==10.1.0
imagehash==4.3.1
requests==2.31.0
python-multipart==0.0.6